- fixed weather CORS issue and with it turn-taking inside the game - cleaned up obsolete files and moved services to better places
