+++ backend/weather/src/main/java/ch/pokino/weather/WeatherController.java
@@ -5,6 +5,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -17,6 +18,7 @@ import java.time.temporal.ChronoUnit;
 
 import static ch.pokino.weather.PokinoWeather.randomWeatherResponse;
 
+@CrossOrigin
 @RestController
 @RequestMapping("/weather")
 public class WeatherController {
