+++ frontend/pokino-app/src/app/model/app-routing.module.ts
@@ -3,16 +3,13 @@ import { RouterModule, Routes } from '@angular/router';
 import { RenderComponent } from '../view/render/render.component';
 
 import {MainMenuComponent} from '../view/main-menu.component';
-import {AppComponent} from '../view/app.component';
 import {LeaderboardComponent} from '../view/leaderboard.component';
-import {WebsocketTestComponent} from "../view/websocket-adapter/websocket-test.component";
 
 const routes: Routes = [
     {path: '', component: MainMenuComponent},
     {path: 'mainMenu', component: MainMenuComponent},
     {path: 'gameScreen', component: RenderComponent},
-    {path: 'leaderboard', component: LeaderboardComponent},
-    {path: 'websocket-test', component: WebsocketTestComponent}
+    {path: 'leaderboard', component: LeaderboardComponent}
 ];
 
 @NgModule({
