--- frontend/pokino-app/src/app/view/websocket-adapter/websocket-test.component.css
@@ -1,9 +1,0 @@
-input {
-    width: 79%;
-    margin-bottom: 1%;
-    font-size: 20px;
-}
-
-.message-container {
-    padding-left: 5%;
-}
